This is not the massive commit you were looking for
 ...did that not work? Thanks for the low Midi-chlorian count MOM!

 Whats new?

 -Redux
 -Refactored structure of site
 -New styling

 Redux
 -----------------------

 I didn't "need" it seeing as how this isn't a super intensive application with too many actions or sharing of state.
 I'd been looking into learning redux and struggling with wrapping my head around everything so I figured implementing it into an application
 that I'm comfortable with would be a good way to get the hang of it. Sure enough I was right.

 Reducers:
    - veggies: handles grabbing and updating inventory count
    - manager: handles logging in users
    - comments: handles grabbing and toggling "archived" state of a comment
    - sales: handles grabbing, presenting [chart.js], updating sales/expenses

Refactored structure of site
-----------------------

In the last iteration of this project the public and private site were mixed. The /dad route would simply render in a log in form within
the public site and upon logging in would allow a user to manipulate the inventory count of a product.
Now things stand alone. The default route takes users to the public site which acts as a website with information on the produce stand
and our current inventory count. The /dad route acts as it's own separate application. Users are presented with a log in form which allows them through to the application itself.
Once they're within the application they have access to three routes

    / [default]
    /comments
    /sales

/ [default]:
    This route allows users to view and manipulate produce. The main actions being incrementing and decrementing the inventory count and removing a product from being "in season" [out of season products won't be shown to the public]

/comments:
    Allows users to view comments left by customers on the public site. You can archive a comments, middleware on the backend filters out archived comments. You can also respond to comments with a simple "mailto:" attribute to an <a>.

/sales:
    This is where users will submit daily expense and sales data. A widget displays the daily totals for sales and expense as well as the net for the day.
    Chart.js is used to display the data on a time series line graph. Users can view the data in time periods of This Week, Month, or Year.

New Styling
-----------------------

...It's black and white now. That's about it.
Oh and i made the icon for the produce cards. NBD I now sorta know how to use Illustrator.